Ast.Env

Documentation

data Env

The Env data type represents an environment that stores mappings of variable names to their values and function names to their definitions.

Constructors

Env

Fields

fn :: Map String Expr

var :: Map String Expr

emptyEnv :: Env

Creates an empty environment with no variables or functions. emptyEnv returns an Env instance with empty mappings for both variables and functions.

insertVar :: String -> Expr -> Env -> Env

Inserts a variable into the environment. insertVar takes as parameters: - String: The name of the variable. - Expr: The value of the variable. - Env: The current environment. And returns a new Env with the variable added to the var mapping.

insertFn :: String -> Expr -> Env -> Env

Inserts a function into the environment. insertFn takes as parameters: - String: The name of the function. - Expr: The definition of the function (usually a lambda expression). - Env: The current environment. and returns a new Env with the function added to the fn mapping.

lookupVar :: String -> Env -> Maybe Expr

Looks up a variable in the environment by its name. lookupVar takes as parameters: - String: The name of the variable. - Env: The current environment. and returns: - `Just Expr`: The value of the variable if it exists in the var mapping. - Nothing: If the variable does not exist in the environment.

lookupFn :: String -> Env -> Maybe Expr

Looks up a function in the environment by its name. lookupFn takes as parameters: - String: The name of the function. - Env: The current environment. and returns: - `Just Expr`: The definition of the function if it exists in the fn mapping. - Nothing: If the function does not exist in the environment.