Ast.Parser.PreProcessor.Define

Documentation

parseDefines :: Parser String

Parses define directives. Supports conditional logic based on environment variables. Returns the processed string after applying defines.

parseDefine :: Parser String

Parses a `?defined` directive. The directive checks if an environment variable is defined. If the variable exists, the thenBlock is returned. Otherwise, the elseBlock (if present) is returned.

parseSet :: Parser String

Parses a `?set` directive. The directive sets an environment variable to an empty value.