Ast.Parser.Program

Documentation

parseProgram :: String -> Parser Program

Parses a program by first preprocessing the source file and then parsing its components. Components can be type definitions or expressions. Returns an Program containing the parsed global expressions, types, and the source file name.

parseTypeDefinition :: Parser Program

Parses a type definition and wraps it in an Program structure. Returns a program with the parsed type definition as a global type.

parseExpr :: Parser Program

Parses an expression and wraps it in an Program structure. Returns a program with the parsed expression as a global expression.

globalExpr :: Expr -> (String, Expr)

Converts a parsed expression into a global expression. Returns a tuple containing the name of the global expression and the expression itself.

globalType :: Type -> (String, Type)

Converts a parsed type into a global type. Returns a tuple containing the name of the global type and the type itself.