Ast.Parser.State

Documentation

type TypeState = [(String, Type)]

Represents the state of type definitions in the parser.

type VarState = [(String, Type)]

Represents the state of variable definitions in the parser.

data PreProcessorState

Tracks preprocessor-specific information, such as visited imports and import depth.

Constructors

PreProcessorState

Fields

visitedImports :: Set String

importDepth :: Int

type DeferState = [[Expr]]

Represents the stack of deferred expressions, grouped by scope.

data ParserState

The complete state of the parser, including types, variables, preprocessor state, and deferred expressions.

Constructors

ParserState

Fields

typeState :: TypeState

varState :: VarState

preProcessorState :: PreProcessorState

deferState :: DeferState

parserState :: ParserState

The initial state of the parser.

insertType :: String -> Type -> ParserState -> ParserState

Inserts a custom type into the environment. If the type already exists, it overwrites it.

lookupType :: String -> ParserState -> Maybe Type

Looks up a custom type in the environment by its name.

insertVar :: String -> Type -> ParserState -> ParserState

Inserts a variable into the environment. If the variable already exists, it overwrites it.

lookupVar :: String -> ParserState -> Maybe Type

Looks up a variable in the environment by its name. Returns Nothing if the variable is not found.

insertImport :: String -> ParserState -> ParserState

Marks an import as visited by adding it to the set of visited imports.

lookupImport :: String -> ParserState -> Bool

Checks if a file has already been imported. Returns True if the file is in the set of visited imports.

setImportDepth :: Int -> ParserState -> ParserState

Sets the depth of nested imports in the preprocessor state.

getImportDepth :: ParserState -> Int

Gets the current depth of nested imports from the preprocessor state.

pushDefered :: Expr -> ParserState -> ParserState

Pushes a deferred expression onto the current defer stack.

pushScope :: ParserState -> ParserState

Pushes a new empty array onto the defer state to represent entering a new scope.

popScope :: ParserState -> ([Expr], ParserState)

Pops the top scope expressions stack. Returns the popped stack and the updated ParserState.