Ast.Parser.UnaryOperation

Documentation

preUnaryOperations :: [(String, UnaryOperation)]

Defines unary operator symbols that precede the operand, mapped to their AST representation. Examples: `++x`, `--x`, `!x`.

postUnaryOperations :: [(String, UnaryOperation)]

Defines unary operator symbols that follow the operand, mapped to their AST representation. Examples: `x++`, `x--`.

parseUnaryOperation :: Parser a -> Parser (UnaryOperation, a)

Parses a unary operator and determines whether it appears before or after the operand.

The function uses lookahead to avoid consuming the operand during parsing. - Pre-unary operators are parsed if they appear before the operand. - Post-unary operators are parsed if they appear after the operand.

operandParser: A parser for the operand, used to determine the position of the operator.

parsePreUnaryOperation :: Parser a -> Parser (UnaryOperation, a)

parsePostUnaryOperation :: Parser a -> Parser (UnaryOperation, a)

parseUnaryOperation' :: [(String, UnaryOperation)] -> Parser UnaryOperation