Ast.Parser.Utils

Documentation

type Parser = ParsecT ParseErrorCustom String (State Env)

A type alias for the parser, based on Parsec with Void error type and String input.

data ParseErrorCustom

Constructors

UnknownType String

UndefinedVar String

UndefinedFunction String

sc :: Parser ()

Skips whitespace and comments (starting with %). Ensures proper handling of spacing in parsers.

lexeme :: Parser a -> Parser a

Wraps a parser to consume trailing whitespace, returning the result of the inner parser.

symbol :: String -> Parser String

Parses a specific symbol (e.g., "+", "-") while skipping trailing whitespace.

triedChoice :: [Parser a] -> Parser a

Tries each parser in the list sequentially, allowing backtracking for all but the last parser.

identifier :: Parser String

An identifier in our language syntax