Codegen.Codegen

Documentation

type MonadCodegen m = (MonadIRBuilder m, MonadModuleBuilder m, MonadFix m)

Type alias for the monad stack used for code generation.

toParamType :: String -> (Type, ParameterName)

Converts a parameter name to a pair of type and parameter name. The toParamType function takes a string and returns a pair of type and parameter name. The type is always i32 (32-bit integer), and the parameter name is the input string.

codegen :: AST -> Module

Generates LLVM code for a given abstract syntax tree (AST). The codegen function takes an AST and returns the corresponding LLVM module.

generateTopLevel :: MonadCodegen m => Expr -> m ()

Generates LLVM code for a top-level expression. The generateTopLevel function takes an expression and generates the corresponding LLVM code.

binaryOps :: [(Operation, Operand -> Operand -> MonadIRBuilder m => m Operand)]

Maps binary operators to LLVM instructions.

generateIf :: MonadCodegen m => Expr -> Expr -> Expr -> m Operand

Generates LLVM code for an if expression.

generateOp :: MonadCodegen m => Operation -> Expr -> Expr -> m Operand

Generates LLVM code for a binary operation.

buildGlobaVariable :: MonadCodegen m => Name -> Literal -> m Operand

Generates LLVM code for a literal variable. The buildLiteralVariable function takes the variable name and its value,

buildLambda :: MonadCodegen m => Name -> [String] -> Expr -> m Operand

Generates LLVM code for a lambda expression. The buildLambda function takes a name, a list of parameter names, and a body expression, and returns an LLVM operand representing the lambda function.

generateExpr :: MonadCodegen m => [(String, Operand)] -> Expr -> m Operand

Generates an LLVM operand for an expression. The generateExpr function recursively processes different expression types and generates the corresponding LLVM code.