Codegen.ExprGen.Assembly

Documentation

callInlineAssembly :: MonadIRBuilder m => InlineAssembly -> Type -> [(Operand, [ParameterAttribute])] -> m Operand

Low level function to generate LLVM code for inline assembly. LLVM's IRM module does not provide a function to generate inline assembly so we have to use the IRBuilder directly.

generateAssembly :: (MonadCodegen m, ExprGen Expr) => Expr -> m Operand

Generate LLVM code for assembly expressions.