Common.Primitives

Documentation

type PrimitiveFunc = [Value] -> Eval Value

Type représentant une fonction primitive qui prend une liste de valeurs et retourne une valeur dans le contexte Eval

initPrimitives :: (Expr -> Eval Value) -> [(String, Value)]

Fonction d'initialisation des primitives | Retourne une liste de tuples (nom, fonction) pour toutes les primitives disponibles

foldNumericOp :: (Integer -> Integer -> Integer) -> Integer -> [Value] -> Eval Value

Fonctions utilitaires pour les opérations numériques | Applique une opération binaire sur une liste de valeurs numériques

numericAdd :: PrimitiveFunc

Opérations numériques de base | Addition: prend une liste de nombres et retourne leur somme

numericSub :: PrimitiveFunc

Soustraction: prend une liste de nombres et retourne leur différence

numericMul :: PrimitiveFunc

Multiplication: prend une liste de nombres et retourne leur produit

numericDiv :: PrimitiveFunc

Division: divise deux nombres entiers

numericMod :: PrimitiveFunc

Modulo: retourne le reste de la division de deux nombres

eq :: PrimitiveFunc

Opérations de comparaison de base | Égalité: vérifie si deux valeurs sont égales

lessThan :: PrimitiveFunc

Inférieur: vérifie si le premier nombre est inférieur au second

greaterThan :: PrimitiveFunc

Supérieur: vérifie si le premier nombre est supérieur au second

lessOrEqual :: PrimitiveFunc

Inférieur ou égal: vérifie si le premier nombre est inférieur ou égal au second

greaterOrEqual :: PrimitiveFunc

Supérieur ou égal: vérifie si le premier nombre est supérieur ou égal au second

notFunc :: PrimitiveFunc

Opérations logiques de base | Not: inverse une valeur booléenne

andFunc :: PrimitiveFunc

And: retourne la conjonction de valeurs booléennes

orFunc :: PrimitiveFunc

Or: retourne la disjonction de valeurs booléennes

stringAppend :: PrimitiveFunc

Opérations sur les chaînes de base | Concaténation: combine plusieurs chaînes de caractères

stringLength :: PrimitiveFunc

Longueur: retourne la longueur d'une chaîne

listFunc :: PrimitiveFunc

Primitives pour les opérations sur les listes | List: crée une nouvelle liste

carFunc :: PrimitiveFunc

Car: retourne le premier élément d'une liste

cdrFunc :: PrimitiveFunc

Cdr: retourne tous les éléments sauf le premier

consFunc :: PrimitiveFunc

Cons: ajoute un élément au début d'une liste

appendFunc :: PrimitiveFunc

Append: concatène deux listes

lengthFunc :: PrimitiveFunc

Length: retourne la longueur d'une liste

nullFunc :: PrimitiveFunc

Null?: vérifie si une liste est vide

isListFunc :: PrimitiveFunc

List?: vérifie si une valeur est une liste

isSymbolFunc :: PrimitiveFunc

Symbol?: vérifie si une valeur est un symbole

factFunc :: PrimitiveFunc

Primitives pour les calculs spéciaux | Fact: calcule la factorielle d'un nombre

diffFunc :: PrimitiveFunc

Diff: vérifie si deux valeurs sont différentes

makeListFunc :: PrimitiveFunc

Make-list: crée une liste de n éléments identiques

setCarFunc :: PrimitiveFunc

Set-car!: modifie le premier élément d'une liste (non autorisé)