EnvStoreRetrieve

Description

Env type export

data Env

Constructors

DefinedExpr Identifier Expr

Variable Identifier (Either Integer Bool)

Retrieving in Env

getKey :: Expr -> Identifier

Takes an Expr as parameter, returns an Identifier.

This function takes an expression as a parameter and retrieve the Identifier. e.g.: Var "foo" -> "foo"

It is used when resolving variables bindings.

lookupDefined :: [Env] -> Expr -> Maybe Expr

Takes a List of Env and an Expr as parameters, returns a Maybe Expr.

this functions tries to find the given expression in the user's defined expressions, whether a defined function or a bound variable. Returns the Expression bound to the variable.

checkList :: Expr -> [Identifier] -> Maybe Identifier

Takes an Expr and a List of Identifier as parameters, returns a Maybe Identifier.

This function takes a list of keywords as parameters and tries to find a correspondance with the given expression.

This function is used to know whether a call is on a builtin or not.

checkCallToken :: [Env] -> Expr -> [Expr] -> Maybe Expr

Takes a List of Env an Expr, and a List of Expr as parameters returns a Maybe Expr

This function checks if the passed expression is a call to an existing function, Either a Builtin or an User defined function. In the case of a builtin, verify param and execute right after. Else returning the body of the defined call.

This is used to reduce a Call or to retrieve a defined function body, to evaluate later (See Interpretor.tryEvalLambda, called by reduceExpr, after verifying function call).

Convert

expToInt :: [Expr] -> Maybe [Integer]

Takes a List of Expr as parameter, returns a Maybe List of Integer.

This function takes an expression list as parameter, and tries to convert it in an Integer list.

This is used to convert and verify that given params to a builtin is a list of Integers values.