Format

Description

Takes some context and builds a custom, nicely formatted informational message.

Logging messages

error :: String

Returns a pretty formatted error message.

warning :: String

Returns a pretty formatted warning message.

Formating functions

fString :: String -> String -> (Int, Int) -> Int -> String -> String

Takes a String, a String, a (Int, Int) and a String as parameters and creates a formatted error message from them.

The first String is the formatted error type, the second String represents the current stream of bytes to display (giving the user some context), the tuple represents the line and column at which the error was detected, the integer is the token's size and the second String is the informational message to display.

fError :: String -> (Int, Int) -> Int -> String -> String

Takes a String, a (Int, Int) and a String as parameters and creates a formatted error message from them.

The first String represents the current stream of bytes to display (giving the user some context), the tuple represents the line and column at which the error was detected, the integer is the token's size and the second String is the informational message to display.

fWarn :: String -> (Int, Int) -> Int -> String -> String

Takes a String, a (Int, Int) and a String as parameters and creates a formatted warning message from them.

The first String represents the current stream of bytes to display (giving the user some context), the tuple represents the line and column at which the warning was detected, the integer is the token's size and the second String is the informational message to display.