Interpreter.Main

Documentation

data Language

Type énumérant les langages supportés par l'interpréteur

Constructors

Python

Lisp

showMenu :: IO Language

Affiche le menu principal et retourne le langage choisi par l'utilisateur

languageMenu :: Language -> IO ()

Affiche et gère le menu spécifique à un langage donné

menuLoop :: IO ()

Boucle infinie gérant l'affichage des menus

handleArgs :: [String] -> IO ()

Gère les arguments passés en ligne de commande pour lancer l'interpréteur | dans différents modes (REPL, exécution de fichier, compilation)

getLanguageFromExt :: String -> Language

Détermine le langage en fonction de l'extension du fichier

showHelp :: IO ()

Affiche l'aide

showVersion :: IO ()

Affiche la version

main :: IO ()

Point d'entrée principal du programme | Configure la sortie sans buffer et lance soit le menu interactif | soit le traitement des arguments selon le cas