Interpretor

Description

and evaluate expressions.

License : MIT Maintainers : maxence.pierreepitech.eu, nathan.flachatepitech.eu

Takes a List of AST and tries to evaluate expressions.

Ast is an abstract way to represent the whole code-base effectively making it easier to interpret and compute

evaluation

interpret :: [Ast] -> [Env] -> Maybe Ast

Takes a List of Ast, a List of Env, returns a Maybe Ast

This is the main Function of the Interpretor, where we can find function definition, as well as function call, or just expression we need to evaluate.

Returns the last evaluation.

Note that an empty Ast returns Nothing, for more explanation on this phenomenon : try compiling an empty file with gcc (dumass)

reduceExpr :: [Env] -> Expr -> Maybe Expr

Takes a List of Env and an Expr returns an Expr

This function is the Core of the interpretor, it is used recursively to effectively reduce the given expression to its minimum, Either an Integer, a Boolean, Or Nothing.

tryEvalLambda :: [Env] -> Expr -> [Expr] -> Maybe Expr

Takes a list of Env, an Expr and a list of Expr, returns a Maybe Expr.

This function tries to parse and evaluate a lambda. /! (May be reused when implementing 'named functions') binds lambda's parameters to given values, if correct, reduce body of the lambda

createLocalEnv :: [Identifier] -> [Expr] -> [Env]

Takes a list of Identifier and an another list of Expr, returns a Maybe '[Env]'

This function tries to mimic variable definition in a case of a function declaration, binding given identifiers to their values.

This function is used when detecting a call to a function with parameters