Language.Python.Compiler

Documentation

compilePython :: String -> Either String [VMInstruction]

Prend du code source Python en entrée et le compile en instructions pour la machine virtuelle. | En cas d'erreur de parsing, retourne l'erreur sous forme de chaîne. | En cas de succès, retourne la liste des instructions compilées.

serialize :: [VMInstruction] -> ByteString

Convertit une liste d'instructions de la machine virtuelle en ByteString | pour pouvoir les sauvegarder dans un fichier ou les transmettre sur le réseau

deserialize :: ByteString -> Either String [VMInstruction]

Convertit un ByteString en liste d'instructions de la machine virtuelle | Utilisé pour charger des instructions précédemment sérialisées