Language.Python.VM

Documentation

executeVM :: [VMInstruction] -> IO (Either String [PyValue])

Execute les instructions de la machine virtuelle Python Retourne soit une erreur soit la liste des valeurs produites

deserialize :: ByteString -> Either String [VMInstruction]

Convertit un ByteString en liste d'instructions de la machine virtuelle | Utilisé pour charger des instructions précédemment sérialisées

data VMState

Constructors

VMState

Fields

globalEnv :: PyEnv

frames :: [Frame]

currentPC :: Int

output :: [String]

data VMInstruction

Constructors

LOAD_CONST PyValue

LOAD_VAR String

STORE_VAR String

BINARY_OP String

UNARY_OP String

CALL String Int

RETURN

JUMP Int

JUMP_IF_FALSE Int

MAKE_LIST Int

MAKE_FUNCTION

POP