Language.Python.VM.Executor

Documentation

executeVM :: [VMInstruction] -> IO (Either String [PyValue])

Execute les instructions de la machine virtuelle Python Retourne soit une erreur soit la liste des valeurs produites

executeInstructions :: [VMInstruction] -> StateT VMState (ExceptT String IO) VMResult

Exécute une séquence d'instructions et gère les retours de fonction

getCurrentFrame :: StateT VMState (ExceptT String IO) Frame

Récupère la frame courante