Lexer

Description

Takes a stream of bytes and performs some checkups, based on a BNF grammar, to extract every Token from it.

If an unexpected character is found or a multi line comment block is not closed, the lexer function returns an error message using fError function.

Lexical analysis

lexer :: Stream -> Either String [(Token, (Int, Int))]

Takes a Stream as parameter and returns a Either String [(Token, (Int, Int))].

On success, this function returns a list of (Token, Int) in which each token has a tuple of integers representing the token's line and column position.

On failure, this function returns a pretty formatted error message.

lexerWrapper :: Stream -> Stream -> (Int, Int) -> Either String [(Token, (Int, Int))]

Takes a Stream and a (Int, Int) as parameters and returns a Either String [(Token, (Int, Int))].

Note that the input Int tuple represents the stream's location starting index for both line and column.

On success, this function returns a list of (Token, Int) in which each token has a tuple of integers representing the token's line and column position.

On failure, this function returns a pretty formatted error message.

Tokens parsing

parseBooleanConstant :: Stream -> Maybe (Token, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Token, Int, Stream) if the stream starts with a BoolLiteral.

On success, this function returns a tuple made of the parsed BoolLiteral, the Lexeme length and the input stream stripped of the parsed Token.

parseKeyword :: Stream -> Maybe (Token, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Token, Int, Stream) if the stream starts with a Token.

On success, this function returns a tuple made of the parsed Token, the Lexeme length and the input stream stripped of the parsed Token.

hParseKeyword :: (Token, Int, Stream) -> Maybe (Token, Int, Stream)

Takes a (Token, Int, Stream) as parameter and returns a Maybe (Token, Int, Stream) if the cut stream starts with a space character.

parseKeyword's helper function that checks if the very next character after the token is a space (to properly create a Keyword and not an Identifier) or a Colon in case of function's parameter type.

parseLiteral :: Stream -> Maybe (Token, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Token, Int, Stream) if the stream starts with a Token.

On success, this function returns a tuple made of the parsed Token, the Lexeme length and the input stream stripped of the parsed Token.

parsePunctuator :: Stream -> Maybe (Token, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Token, Int, Stream) if the stream starts with a Token.

On success, this function returns a tuple made of the parsed Token, the Lexeme length and the input stream stripped of the parsed Token.

Lexemes parsing

parseMultiLineComment :: Stream -> Stream -> (Int, Int) -> (Int, Int) -> Either String (Stream, (Int, Int))

Takes a 2 Stream (stream's start and stream's current position) and 2 (Int, Int) (starting position and current position) as parameters and returns a Either String (Stream, (Int, Int)).

On success, this function returns a tuple made of the parsed comment block, enclosed in `\*` and `*/` and the new Stream's current position. On failure, this function returns a pretty formatted error message.

parseStringLiteral :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <string-literal>.

This function is a wrapper around parseSCharSequence that checks whether the string literal starts and ends with the `"` character. As such, it returns the same components and increases the total length by 2, taking into account both `"` characters.

parseSCharSequence :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <s-char-sequence>.

On success, this function returns a tuple made of the parsed string literal, the string length and the input stream stripped of the parsed string literal.

This function differs from parseCharacterConstant in that it accepts one or many characters and is surrounded by `"` characters.

parseSChar :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <s-char>.

On success, this function returns a tuple made of the parsed character, the character length (which is always 1) and the input stream stripped of the parsed character.

This function differs from parseChar as it fails if the character `"` is not escaped.

parseDecimalConstant :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <decimal-constant>.

On success, this function returns a tuple made of the decimal string representation, the parsed integer length and the input stream stripped of the parsed integer.

parseFloatingConstant :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <floating-constant>.

On success, this function returns a tuple made of the float string representation, the parsed float length (including the dot separator) and the input stream stripped of the parsed float.

parseDigitSequence :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <digit-sequence>.

On success, this function returns a tuple made of the digit string representation, the parsed digit length and the input stream stripped of the parsed digit.

hParseDigitSequence :: Stream -> Maybe (Lexeme, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Int, Stream) if the stream starts with a <digit-sequence>.

parseDigitSequence's helper function that if the current stream a a digit sequence.

parseCharacterConstant :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <character-constant>.

On success, this function returns a tuple made of the parsed string literal, the string length and the input stream stripped of the parsed string literal.

This function differs from parseSCharSequence in that it accepts only one character and is surrounded by `'` characters.

parseChar :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <s-char>.

On success, this function returns a tuple made of the parsed character, the character length (which is always 1) and the input stream stripped of the parsed character.

This function differs from parseSChar as it fails if the character `'` is not escaped.

parseEscapeSequence :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <escape-sequence>.

On success, this function returns a tuple made of the parsed character, the character length (which is always 1) and the input stream stripped of the parsed character.

parseIdentifier :: Stream -> Maybe (Identifier, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <identifier>.

On success, this function returns a tuple made of the parsed identifier, the identifier length and the input stream stripped of the identifier.

parseNonDigit :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <nondigit>.

On success, this function returns a tuple made of the parsed character, the character length (which is always 1) and the input stream stripped of the parsed character.

parseDigit :: Stream -> Maybe (Char, Int, Stream)

Takes a Stream as parameter and returns a Maybe (Char, Int, Stream) if the stream starts with a <digit>.

On success, this function returns a tuple made of the parsed digit (as a Char), the digit length (which is always 1) and the input stream stripped of the parsed digit.