Parser

Description

Takes a list of Token and do an syntaxic analysis, based on a BNF grammar, to build an Decl list representing the program Abstract Syntax Tree.

If an unexpected token is found or the syntax is invalid, the parser function returns an pretty :) error message.

Documentation

parser :: [SingleToken] -> Either String [Decl]

Takes a [SingleToken] as parameter and returns a Either String [Decl].

On success, this function returns a list of Decl representing the complete AST of the program.

On failure, this function returns a pretty formatted error message.

parseDecl :: Parser [Decl]

Takes a Parser SingleToken list as parameter and returns a Either String ([Decl], [SingleToken]).

On success, this function returns a tuple made of the parsed declaration list and the remaining tokens.

On failure, this function returns a pretty formatted error message.

parseTopLevel :: Parser Decl

Takes a Parser SingleToken list as parameter and returns a Either String (Decl, [SingleToken]).

On success, this function returns a tuple made of the parsed top-level declaration (either a function or variable declaration) and the remaining tokens.

On failure, this function returns a pretty formatted error message.

parseFunctionDecl :: Parser Decl

Takes a Parser SingleToken list as parameter and returns a Either String (Decl, [SingleToken]).

On success, this function returns a tuple made of the parsed function decl.

On failure, this function returns a pretty formatted error message.

parseVarDecl :: Decl -> Parser Decl

Takes a Parser SingleToken list as parameter and returns a Either String (Decl, [SingleToken]).

On success, this function returns a tuple made of the parsed variable declaration as VarDecl and the remaining tokens.

On failure, this function returns a pretty formatted error message if invalid syntax or a semicolon is missing.