ParserStmt

Description

Handles the parsing of statements including if/else, while loops, for loops, return statements, and compound statements.

If an unexpected token is found or if the syntax is invalid, the parsing functions return an pretty formatted error message.

Documentation

parseCompoundStmt :: Decl -> Parser CompoundStmt

Takes an Decl and a '[SingleToken]' as parameter and returns a Either String CompoundStmt.

On success, this function returns a Stmt.

On failure, this function returns a pretty formatted message error.

This function is used to parse a compound statement, (e.g.: a function's body, a for's body).