Tokens

Description

Sets different categories to which each token belongs. The tokens are based on the C programming language.

For further informations, refer to the Microsoft C lexical grammar definition from which this definition is based.

BNF definition

For the full rizz lexical syntax definition, see the BNF definition here.

Basic type

type Stream = String

Defines Stream type as a string representing a finite byte array.

type Lexeme = String

Defines Lexeme type as a string representing a way to identify a specific Token.

type Identifier = String

Defines Identifier type as a string representing a variable.

Tokens kind

data Token

Defines Token which is bound by its corresponding representation in the bytes Stream.

It can only be of one type and serves as a way to represent a series of bytes in a more abstract way.

Constructors

Keyword Keyword
Keyword definition.

Identifier Identifier
Identifier definition.

Literal Literal
Literal definition.

Punctuator Punctuator
Punctuator definition.

data Keyword

Defines Token which are reserved words used by the language.

Constructors

Bool
Boolean type keyword, expressed in rizz code as `Bool`.

Char
Character type keyword, expressed in rizz code as `Char`.

Int
Integer type keyword, expressed in rizz code as `Int`.

Float
Float type keyword, expressed in rizz code as `Float`.

Double
Double type keyword, expressed in rizz code as `Double`.

Fn
Function declaration keyword, expressed in rizz code as `fn`.

If
If keyword, expressed in rizz code as `if`.

Else
Else keyword, expressed in rizz code as `else`.

While
While loop type keyword, expressed in rizz code as `while`.

For
For loop keyword, expressed in rizz code as `for`.

Foreach
Foreach loop keyword, expressed in rizz code as `foreach`.

Ret
Return (from function) keyword, expressed in rizz code as `ret`.

Struct
Struct keyword, expressed in rizz code as `struct Foo = {...}`.

data Literal

Defines Token which is a way to represent a constant value not bound to a variable.

Constructors

BoolLiteral Bool
Boolean literal, expressed in rizz code as `True` or `False`.

CharLiteral Char
Character literal, expressed in rizz code like `'x'`.

IntLiteral Integer
Integer literal, expressed in rizz code like `42`.

FloatLiteral Float
Single floating point literal, expressed in rizz code like `3.14`.

ListLiteral [Literal]
List literal, expressed in rizz code like `[1, 2, 3]`.

data SBracket

Defines SBracket representing both opening and closing square brackets `[]`.

Constructors

OpenSBracket
Opening square bracket, expressed in rizz code as `[`.

CloseSBracket
Closing square bracket, expressed in rizz code as `]`.

data RBracket

Defines RBracket representing both opening and closing round brackets `()`.

Constructors

OpenRBracket
Opening round bracket, expressed in rizz code as `(`.

CloseRBracket
Closing round bracket, expressed in rizz code as `)`.

data CBracket

Defines RBracket representing both opening and closing curly brackets `{}`.

Constructors

OpenCBracket
Opening curly bracket, expressed in rizz code as `{`.

CloseCBracket
Closing curly bracket, expressed in rizz code as `}`.

data UnaryOp

Defines UnaryOp representing both increment and decrement unary operators.

Constructors

IdentIncrement
Unary increment operator, expressed in rizz code as `++`.

IdentDecrement
Unary decrement operator, expressed in rizz code as `--`.

data BinaryOp

Defines BinaryOp representing arithmetic and equality operators.

Constructors

Mul
Multiplication operator, expressed in rizz code as `*`.

Add
Addition operator, expressed in rizz code as `+`.

Sub
Subtraction operator, expressed in rizz code as `-`.

Div
Division operator, expressed in rizz code as `/`.

Mod
Modulo operator, expressed in rizz code as `%`.

Lt
Less than operator, expressed in rizz code as `<`.

Gt
Greater than operator, expressed in rizz code as `>`.

LEq
Less than or equal to operator, expressed in rizz code as `<=`.

GEq
Greater than or equal to operator, expressed in rizz code as `>=`.

Eq
Equal operator, expressed in rizz code as `==`.

NEq
Not equal operator, expressed in rizz code as `!=`.

And
And operator, expressed in rizz code as `&&`.

Or
Or operator, expressed in rizz code as `||`.

data AssignOp

Defines AssignOp representing assignment operators with arithmetic operation on it.

Constructors

Equal
Assignment operator, expressed in rizz code as `=`.

MulEqual
Multiplication and assignment operator, expressed in rizz code as `*=`.

DivEqual
Division and assignment operator, expressed in rizz code as `/=`.

ModEqual
Modulo and assignment operator, expressed in rizz code as `%=`.

AddEqual
Addition and assignment operator, expressed in rizz code as `+=`.

SubEqual
Subtraction and assignment operator, expressed in rizz code as `-=`.

data Punctuator

Defines Token which is a large category containing non keywords, non literals and non identifiers tokens.

Constructors

SBracket SBracket
Both square brackets, used to index lists.

RBracket RBracket
Both curly brackets, used to declare or call functions.

CBracket CBracket
Both curly brackets, used to define a scope.

Dot
Standalone dot Token , used to declare floating point constants and expressed in rizz code as `.`.

Arrow
Standalone arrow Token, used to specify the return type of a function and expressed in rizz code as `+=`.

UnaryOp UnaryOp
Both unary operators.

BinaryOp BinaryOp
all of binary operators.

Colon
Standalone colon Token, used to specify the type of an Identifier and expressed in rizz code as `:`.

Semicolon
Standalone semicolon Token, used to specify the end of a statement and expressed in rizz code as `;`.

Comma
Standalone comma Token, used to space functions' arguments and expressed in rizz code as `,`.

QMark
Standalone question mark Token, used to mark a ternary condition and expressed in rizz code like `True ? {} : {}`.

AssignOp AssignOp
All of assignment operators.