Ast.Parser

Documentation

parse :: String -> String -> Either String AST

Parses a string into an abstract syntax tree (AST). The parse function takes a filename String, and an input String as a parameter and returns either an AST or an error message.

data ParseErrorCustom

Custom error type for the parser, representing specific cases of invalid syntax.

Constructors

UndefinedLambdaReference String

ReservedKeywordUsed String

UndefinedVarReference String

InvalidVarName String