Codegen.Codegen

Documentation

codegen :: AST -> Module

Generates LLVM code for a given abstract syntax tree (AST). The codegen function takes an AST and returns the corresponding LLVM module.

generateExpr :: (MonadIRBuilder m, MonadFix m) => Expr -> m Operand

Generates an LLVM operand for an expression. The generateExpr function recursively processes different expression types and generates the corresponding LLVM code.