Codegen.Codegen

Documentation

type CodegenState = [(String, Operand)]

Type alias for the code generation state.

data CodegenError

Type for codegen errors

Constructors

UnsupportedTopLevel Expr

UnsupportedOperator Operation

UnsupportedLiteral Literal

UnsupportedGlobalVar Literal

UnsupportedLocalVar Literal

UnsupportedDefinition Expr

VariableNotFound String

UnsupportedFunctionCall String

type MonadCodegen m = (MonadIRBuilder m, MonadModuleBuilder m, MonadFix m, MonadState CodegenState m, MonadError CodegenError m)

Type alias for the monad stack used for code generation.

getVarBinding :: MonadCodegen m => String -> m (Maybe Operand)

Helper functions to manage state

addVarBinding :: MonadCodegen m => String -> Operand -> m ()

Adds a variable binding to the state.

toParamType :: String -> (Type, ParameterName)

Converts a parameter name to a pair of type and parameter name. The toParamType function takes a string and returns a pair of type and parameter name. The type is always i64 (64-bit integer), and the parameter name is the input string.

codegen :: AST -> Either CodegenError Module

Generates LLVM code for a given abstract syntax tree (AST).

generateTopLevel :: MonadCodegen m => Expr -> m ()

Generates LLVM code for a top-level expression. The generateTopLevel function takes an expression and generates the corresponding LLVM code.

binaryOps :: [(Operation, Operand -> Operand -> MonadIRBuilder m => m Operand)]

Maps binary operators to LLVM instructions.

generateIf :: MonadCodegen m => Expr -> Expr -> Expr -> m Operand

Generates LLVM code for an if expression.

generateOp :: MonadCodegen m => Operation -> Expr -> Expr -> m Operand

Generates LLVM code for a binary operation.

buildGlobaVariable :: MonadCodegen m => Name -> Literal -> m Operand

Generates LLVM code for a literal variable. The buildLiteralVariable function takes the variable name and its value,

buildLambda :: MonadCodegen m => Name -> [String] -> Expr -> m Operand

Generates LLVM code for a lambda expression. The buildLambda function takes a name, a list of parameter names, and a body expression, and returns an LLVM operand representing the lambda function.

buildFunction :: MonadCodegen m => Name -> [String] -> Expr -> m Operand

Generates LLVM code for a lambda expression. The buildFunction will build both a lambda and a named function. The function generated will just call the lambda function and return the result.

generateExpr :: MonadCodegen m => Expr -> m Operand

Generates an LLVM operand for an expression. The generateExpr function recursively processes different expression types and generates the corresponding LLVM code.

generateLiteral :: MonadCodegen m => Literal -> m Operand

Generates an LLVM operand for a literal. The generateLiteral function takes a literal and returns the corresponding LLVM operand.

generateCall :: MonadCodegen m => Expr -> Expr -> m Operand

Generates an LLVM operand for a function call. The generateCall function takes a function expression and a list of argument expressions, and returns the corresponding LLVM operand.

generateVar :: MonadCodegen m => String -> m Operand

Generates an LLVM operand for a variable. The generateVar function takes a variable name and returns the corresponding LLVM operand.

generateDefine :: MonadCodegen m => String -> Expr -> m Operand

Generates an LLVM operand for a definition. The generateDefine function takes a variable name and an expression, and returns the corresponding LLVM operand.

generateLambda :: MonadCodegen m => [String] -> Expr -> m Operand

Generates an LLVM operand for a lambda expression. The generateLambda function takes a list of parameter names and a list of body expressions, and returns the corresponding LLVM operand.

generateSeq :: MonadCodegen m => [Expr] -> m Operand

Generates an LLVM operand for a sequence of expressions. The generateSeq function takes a list of expressions and returns the corresponding LLVM operand.