Lexer

Description

Takes a stream of bytes and performs some checkups to extract every Token from it, based on pre-defined Lexemes.

Token is a way to break down a series of bytes into an individual, small piece that represent something in the global context of the parsed stream.

Lexemes, however, are a list of predetermined Lexeme that help to determine a tokens among a bytes stream.

Basic types

type Stream = String

Defines Stream type as a string which represents a finite byte stream.

type Lexeme = String

Defines Lexeme type as a string representing a way to identify a specific Token.

type Lexemes = [Lexeme]

Defines Lexemes type as a list of Lexeme, covering a set of patterns to match for a Token identification.

data Token

Defines a Token which is bound by its corresponding representation in the bytes Stream.

It can only be of one type and serves as a way to represent a series of bytes in a more abstract way.

Constructors

Keyword Lexeme
Keyword is any lexeme defined in the keywords list. It represents a reserved word used by LISP language.

Identifier Lexeme
Identifier is a variable, function or lambda name built when no other Token can be built from current Stream and that is not a Constant nor a Boolean.

Operator Lexeme
Operator is any lexeme defined in the operators list. It represents auto-bound functions used for arithmetic and comparison operations.

Delimiter Lexeme
Delimiter is any lexeme defined in delimiters list. It represents a character used to separate some other tokens.

Constant Integer
Constant is any signed integerer representing a constant value.

Boolean Bool
Boolean is either True or False and is represented by either '#t' for True and '#f' for False values.

Lexemes

keywords :: Lexemes

Defines keywords as define, lambda and if patterns. Note that an Identifier can start with any keyword as long as their are no spaces between the keyword and the identifier.

operators :: Lexemes

Defines operators as +, -, *, /, <, > and eq? patterns.

delimiters :: Lexemes

Defines delimiters as ( and ) patterns.

Parsing

parseAnyToken :: Stream -> Maybe (Token, Stream)

Takes a Stream as parameter and returns a Maybe (Token, Stream).

This function tries to parse any Token from the current Stream (either from keywords, operators or delimiters) and returns Nothing if no valid Token is found.

parseToken :: Stream -> Lexemes -> Maybe (Lexeme, Stream)

Takes a Stream and Lexemes as parameters and returns a Maybe (Lexeme, Stream).

This function takes a Lexemes and checks if any of its Lexeme match the first bytes of the given Stream. If none of the Lexemes were found at the beginning of the stream, returns Nothing.

parseIdentifier :: Stream -> Maybe (Lexeme, Stream)

Takes a Stream as parameter and returns a Maybe (Lexeme, Stream).

This functions takes a Stream and tries to parses it as a Token Identifier and returns Nothing if it fails.

parseConstant :: Stream -> Maybe (Token, Stream)

Takes a Stream as parameter and returns a Maybe ('Token, Stream).

This functions takes a Stream and tries to parses it either a Token Constant or Boolean and returns Nothing if it fails.

parseUInteger :: Stream -> Maybe (String, Stream)

Takes a Stream as parameter and returns a Maybe (String, Stream).

This functions takes a Stream and tries to extracts a valid signed Integer and returns Nothing if it fails.

Lexing

lexer :: Stream -> [Token]

Takes a Stream as parameter and returns a list of Token.

This functions parses the given stream and extracts every tokens to a list that is returned.

A special case, not covered by lexemes, are comments, starting with a '\n' that let a user define a comment in its code that should be ignored. However, Lexemes from keywords, operators or delimiters are used to generate the list of Token.