Parser

Description

Takes a List of Token and tries to parse expressions and return it to a List.

Token is a way to break down a series of bytes into an individual, small piece that represent something in the global context of the parsed stream.

List is a list of Ast who is an abstract way to represent the whole code-base effectively making it easier to interpret and compute

Main function

getAST :: [Token] -> IO List

Get a list of Ast by a list of Token

Subfunction

parsor :: [Token] -> Either ErrorT List

Parse all tokens into a Either ErrorT List

parseTopLevel :: Parser Ast

Parse all tokens into a Ast